requirements_txt.txt
streamlit
langchain-community
faiss-cpu
google-generativeai
pypdf
pdf2image
pytesseract
python-dotenv
pillow